requirements.txt
bcrypt
mysql-connector-python
pandas
pydantic[email]
pydantic
pyinstaller
streamlit
